added css for dashboard components
