+++ frontend/src/components/DashboardLayout.css
@@ -1,0 +1,20 @@
+body, .dashboard-container {
+  font-family: 'Inter', sans-serif;
+  background-color: white;
+  max-width: 100vw;
+  margin: auto;
+  padding: 1rem;
+}
+
+.dashboard-main {
+  display: flex;
+  margin-top: 1.5rem;
+  gap: 1.5rem;
+}
+
+.main-content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
