+++ frontend/src/components/Header.css
@@ -1,0 +1,87 @@
+.dashboard-header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  border-radius: 1rem;
+  padding: 0rem 0.5rem;
+}
+
+.masked-header {
+  background-color: #D9D9D9;
+  display: flex;
+  align-items: center;
+  border-radius: 1.5rem;
+  padding: 0.45rem 1.5rem;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+}
+
+.search-bar {
+  flex: 3;
+  max-width: 60vw;
+  margin: 0 1rem;
+  padding: 0.8rem 0.8rem;
+  padding-right: 15rem;
+  border-radius: 9999px;
+  background-color: #E5E5E5;
+  border: none;
+  color: #9CA3AF;
+}
+
+.header-icons {
+  display: flex;
+  gap: 0.75rem;
+}
+
+.icon-circle {
+  width: 2rem;
+  height: 2rem;
+  background-color: #FF6F3C;
+  border-radius: 9999px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  transition: transform 0.2s ease;
+}
+
+.icon-circle:hover {
+  transform: translateY(-3px);
+}
+
+.icon-avatar {
+  width: 3rem;
+  height: 3rem;
+  background-color: #FF6F3C;
+  border-radius: 9999px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  transition: transform 0.3s ease;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: white;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.icon-avatar:hover {
+  transform: scale(1.1);
+}
+
+.popover {
+  position: absolute;
+  top: 110%;
+  right: 0;
+  background-color: white;
+  padding: 0.75rem 1rem;
+  border-radius: 0.5rem;
+  box-shadow: 0 6px 20px rgba(0, 0, 0, 0.1);
+  z-index: 10;
+}
+
+.popover-item {
+  cursor: pointer;
+  color: #374151;
+  font-size: 0.875rem;
+  padding: 0.25rem 0;
+}
+
+.popover-item:hover {
+  color: #FF6F3C;
+}
