+++ frontend/src/components/Header.js
@@ -1,6 +1,8 @@
 import React, { useState, useRef, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { jwtDecode } from "jwt-decode";
+import './Header.css';
+
 
 const Header = () => {
   const navigate = useNavigate();
