+++ frontend/src/pages/assets/icon_freelancer.png
