+++ frontend/src/pages/assets/icon_seller.png
